+++ src/main/java/net/MrGise/floating/helper/CreateBuilderHelper.java
@@ -8,11 +8,9 @@ import com.tterrag.registrate.builders.BlockBuilder;
 import com.tterrag.registrate.util.nullness.NonNullUnaryOperator;
 import net.MrGise.floating.FloatingIslands;
 import net.MrGise.floating.ModIDs;
-import net.MrGise.mmm.MMM;
 import net.MrGise.mmm.block.PortalBlock;
 import net.MrGise.mmm.item.block_item.description.DescriptionPortalBlockItem;
 import net.MrGise.mmm.registry.create.ModCTBehaviour;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraftforge.client.model.generators.ConfiguredModel;
@@ -66,7 +64,7 @@ public class CreateBuilderHelper {
                     }
                 }))
                 .onRegister(connectedTextures(() -> new ModCTBehaviour.PortalCTBehaviour(ctOff.get(), ctOn.get())))
-                .item((b1, properties1) -> new DescriptionPortalBlockItem(b1, properties1, eyeName, MMM.MOD_ID))
+                .item((b1, properties1) -> new DescriptionPortalBlockItem(b1, properties1, eyeName, ModIDs.mmm()))
                 .model((context, provider) ->
                         provider.withExistingParent("connecting_portal_block", mmm("block/connecting_portal_block_on")))
                 .build();