+++ src/main/java/net/MrGise/mmm/network/ModNetwork.java
@@ -1,7 +1,9 @@
 package net.MrGise.mmm.network;
 
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.network.NetworkRegistry;
+import net.minecraftforge.network.PacketDistributor;
 import net.minecraftforge.network.simple.SimpleChannel;
 
 public class ModNetwork {
@@ -24,5 +26,16 @@ public class ModNetwork {
                 SyncAllKnowingPacket::new,
                 SyncAllKnowingPacket::handle
         );
+        CHANNEL.registerMessage(
+                packetId++,
+                OpenScreenPacket.class,
+                (msg, buf) -> buf.writeUtf(msg.screenId),
+                buf -> new OpenScreenPacket(buf.readUtf()),
+                OpenScreenPacket::handle
+        );
+    }
+
+    public static void sendToPlayer(ServerPlayer player, Object message) {
+        CHANNEL.send(PacketDistributor.PLAYER.with(() -> player), message);
     }
 }